college-event-backend/requirements.txt
Flask
pymongo
python-dotenv
flask-cors
requests
fpdf
